fastmcp/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastmcp"
version = "0.1.0"
description = "FastMCP - secure orchestration layer for enterprise agents"
authors = [{name = "FastMCP Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.28",
    "sqlmodel>=0.0.16",
    "pydantic>=2.5",
    "pydantic-settings>=2.2",
    "python-jose[cryptography]>=3.3",
    "cryptography>=41.0",
    "httpx>=0.27",
    "tenacity>=8.2",
    "structlog>=23.2",
    "redis>=5.0",
    "jsonschema>=4.21",
    "python-dotenv>=1.0",
    "opentelemetry-api>=1.23",
    "opentelemetry-sdk>=1.23",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "anyio>=3.7",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
